web-logs/requirements.txt
openai
pinecone
pinecone-client==2.2.1
python-dotenv==1.0.0
